+++ packages/dde-earth/src/core/index.ts
@@ -4,3 +4,5 @@ export * from "./layerItem";
 export * from "./layerManager";
 export * from "./terrainManager";
 export * from "./debug";
+export * from "./event";
+export * from "./pluginManager";