+++ packages/plugin-feature-click/rollup.config.js
@@ -1,0 +1,3 @@
+import { createConfig } from "../../shared/rollup.config.mjs";
+
+export default createConfig();