+++ packages/recommend-plugins/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @dde-earth/recommend-plugins
 
+## 1.3.3
+
+### Patch Changes
+
+- Updated dependencies
+  - dde-earth@1.2.3
+
 ## 1.3.2
 
 ### Patch Changes